+++ changelog.md
@@ -1,5 +1,9 @@
 # Version log
 ## 0.0.7dev
+###Thu  1 Jun 17:30:09 BST 2023
+- issue#27 : hkl to shelx
+  - script
+  - download button
 ###Thu 1 Jun 10:54:33 BST 2023
 - issue#29:bloch rock
   - enable exp/sim rocking curves option
@@ -10,6 +14,7 @@
 - changed the `info.graphs` structure
 - get automatic reflection selection in rock mode when refreshing
 - `clear` flag in `update_refl` callback when clearing table
+
 ## 0.0.6
 ###Wed 31 May 22:10:13 BST 2023
 - issue#29 : bloch rock